Function Payment With Paypal

+++ database/migrations/2018_09_29_075343_create_returs_table.php
@@ -20,8 +20,8 @@ class CreateRetursTable extends Migration
             $table->string('kode_user');
             $table->text('keterangan');
             $table->date('tgl_retur');
-            $table->integer('grandtotal');
-            $table->integer('cashback');
+            $table->float('grandtotal');
+            $table->float('cashback');
             $table->string('status');
             $table->timestamps();
         });
