+++ database/migrations/2018_09_29_075347_create_opsi_returs_table.php
@@ -18,8 +18,8 @@ class CreateOpsiRetursTable extends Migration
             $table->string('kode_retur');
             $table->string('kode_barang');
             $table->integer('qty');
-            $table->integer('harga');
-            $table->integer('subtotal');
+            $table->float('harga');
+            $table->float('subtotal');
             $table->timestamps();
         });
     }
